+++ src/promptflow/tests/sdk_cli_azure_test/unittests/test_flow_operations.py
@@ -1,7 +1,6 @@
 # ---------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
-
 from pathlib import Path
 
 import pytest
@@ -22,19 +21,22 @@ class TestFlowOperations:
 
         flow_source = flow_test_dir / "web_classification/"
         with pytest.raises(UserErrorException, match="Not a valid string"):
-            pf.flows.create_or_update(flow=flow_source, display_name=object())
+            pf.flows.create_or_update(flow=flow_source, display_name=False)
 
         with pytest.raises(UserErrorException, match="Must be one of: standard, evaluation, chat"):
             pf.flows.create_or_update(flow=flow_source, type="unknown")
 
         with pytest.raises(UserErrorException, match="Not a valid string"):
-            pf.flows.create_or_update(flow=flow_source, description=object())
+            pf.flows.create_or_update(flow=flow_source, description=False)
 
         with pytest.raises(UserErrorException, match="Not a valid string"):
-            pf.flows.create_or_update(flow=flow_source, tags={"key": object()})
+            pf.flows.create_or_update(flow=flow_source, tags={"key": False})
 
-        with pytest.raises(UserErrorException, match="Unknown field"):
-            pf.flows.create_or_update(flow=flow_source, random="random")
+    @pytest.mark.usefixtures("enable_logger_propagate")
+    def test_create_flow_with_warnings(self, pf, caplog):
+        flow_source = flow_test_dir / "web_classification/"
+        pf.flows._validate_flow_creation_parameters(source=flow_source, random="random")
+        assert "random: Unknown field" in caplog.text
 
     def test_list_flows_invalid_cases(self, pf):
         with pytest.raises(FlowOperationError, match="'max_results' must be a positive integer"):